+++ npm/vite-dev-server/cypress/e2e/vite-dev-server.cy.ts
@@ -22,7 +22,7 @@ describe('Config options', () => {
     cy.visitApp()
 
     cy.contains('App.cy.jsx').click()
-    cy.get('.passed > .num').should('contain', 1)
+    cy.get('.passed > .num').should('contain', 2)
 
     cy.withCtx(async (ctx) => {
       const config = ctx.lifecycleManager.loadedFullConfig
@@ -47,7 +47,7 @@ describe('Config options', () => {
 
     cy.visitApp()
     cy.contains(specWithWhitespace).click()
-    cy.get('.passed > .num').should('contain', 1)
+    cy.get('.passed > .num').should('contain', 2)
   })
 
   it('supports @cypress/vite-dev-server', () => {
@@ -58,7 +58,7 @@ describe('Config options', () => {
     cy.visitApp()
     cy.contains('App.cy.jsx').click()
     cy.waitForSpecToFinish()
-    cy.get('.passed > .num').should('contain', 1)
+    cy.get('.passed > .num').should('contain', 2)
   })
 
   it('supports viteConfig as an async function', () => {
@@ -69,7 +69,7 @@ describe('Config options', () => {
     cy.visitApp()
     cy.contains('App.cy.jsx').click()
     cy.waitForSpecToFinish()
-    cy.get('.passed > .num').should('contain', 1)
+    cy.get('.passed > .num').should('contain', 2)
     cy.withCtx(async (ctx) => {
       const verifyFile = await ctx.file.readFileInProject('wrote-to-file')